requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.5
crispy-bootstrap5==0.6
decorator==5.1.1
Django==4.0.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
executing==0.8.3
Faker==13.13.0
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
ipython==8.4.0
jedi==0.18.1
marshmallow==3.16.0
matplotlib-inline==0.1.3
mccabe==0.7.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
tzdata==2022.1
wcwidth==0.2.5
webargs==8.1.0
